Fix circular imports on native
